Create a Shell Script

Create a shell script for a data analysis workflow consisting of commands extracted from existing R files.

Usage

shell_script(
  files,
  dir_shell = NULL,
  name_shell = "workflow.sh",
  overwrite = FALSE
)

Arguments

files

A path from dir_shell to a directory with R files that have calls to cmd_assign().

dir_shell

The directory where shell_script() will create the shell script. If no value is supplied, then `shell_script(); creates the shell script the current working directory.

name_shell

The name of the shell script. The default is "workflow.sh".

overwrite

Whether to overwrite an existing file. Default is FALSE.

Value

shell_script() is called for its side effect, which is to create a file. However, shell_script() also returns a string with the text for a shell script.

Details

To create a shell script in the files directory, set files to ".".

To obtain the contents of the shell script without creating a file on disk, creating the file on disk, set name_shell to NULL.

Supplying a value for files is compulsory for shell_script(), but optional for makefile(). The output from shell_script() is generated entirely from files while the output from makefile() also includes some general-purpose Makefile commands.

References

Episodes 1–3 of The Unix Shell Introduction to the command line

Command-Line Programs Introduction to Rscript

See also

Examples

library(fs)

## Create project directory containing
## 'src' and 'out' directories
path_project <- file_temp()
dir_create(path_project)
dir_create(path(path_project, "src"))
dir_create(path(path_project, "out"))

## Put R code file in 'src' directory
writeLines(c("cmd_assign(x = 1, .out = 'out/results.rds')",
             "results <- x + 1",
             "saveRDS(results, file = .out)"),
           con = path(path_project, "src/results.R"))

## Look at directories
dir_tree(path_project)
#> /tmp/RtmpO6a97r/file1ec352cd2570
#> ├── out
#> └── src
#>     └── results.R

## Call 'shell_script()'
shell_script(files = "src",
             dir_shell = path_project)
#> ✔ Extracted call to `cmd_assign()` in src/results.R.

## Look at directories
dir_tree(path_project)
#> /tmp/RtmpO6a97r/file1ec352cd2570
#> ├── out
#> ├── src
#> │   └── results.R
#> └── workflow.sh

## Look at contents of shell script
lines <- readLines(path(path_project, "workflow.sh"))
cat(paste(lines, collapse = "\n"))
#> 
#> Rscript src/results.R \
#>   out/results.rds \
#>   --x=1
#> 

## Get the text of the shell script
## without creating a file on disk
text <- shell_script(files = "src",
                     dir_shell = path_project,
                     name_shell = NULL)
#> ✔ Extracted call to `cmd_assign()` in src/results.R.
cat(text)
#> 
#> Rscript src/results.R \
#>   out/results.rds \
#>   --x=1
#> 

## Clean up
dir_delete(path_project)